Hw04 auth

+++ controllers/contacts/deleteContacts.js
@@ -1,20 +1,20 @@
-const {Contact} = require("../../models/contact");
+const { Contact } = require("../../models/contact");
 const { HttpError } = require("../../helpers");
 
-
 const deleteContact = async (req, res, next) => {
-    try {
-      const { contactId } = req.params;
-      const result = await Contact.findByIdAndRemove(contactId);
-      console.log("result:", result)
-      if (!result) {
-        throw HttpError(404, "Not found");
-      }
-  
-      res.json({message: "Delete success"});
-    } catch (error) {
-      next(error);
+  try {
+    const { contactId } = req.params;
+    const { _id: owner } = req.user;
+    const result = await Contact.findByIdAndRemove({ contactId, owner });
+    console.log("result:", result);
+    if (!result) {
+      throw HttpError(404, "Not found");
     }
-  }
 
-  module.exports = deleteContact;
+    res.json({ message: "Delete success" });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = deleteContact;
